refactor: More reusable table elements to under internal/components/table-fragments
