+++ src/area-chart/elements/area-chart-legend.tsx
@@ -4,8 +4,8 @@ import React, { memo, useMemo } from 'react';
 
 import { AreaChartProps } from '../interfaces';
 import ChartLegend from '../../internal/components/chart-legend';
-import { useSelector } from '../model/async-store';
 import { ChartModel } from '../model';
+import { useSelector } from '../../internal/async-store';
 
 export default memo(AreaChartLegend) as typeof AreaChartLegend;
