+++ src/area-chart/model/interactions-store.ts
@@ -2,8 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 import { ChartModel } from './index';
 import { AreaChartProps } from '../interfaces';
-
-import AsyncStore from './async-store';
+import AsyncStore from '../../internal/async-store';
 
 const initialState = Object.freeze({
   highlightedX: null,
