+++ src/internal/components/table-fragments/sticky-columns/__tests__/use-sticky-columns.test.tsx
@@ -5,9 +5,9 @@ import {
   useStickyColumns,
   StickyColumnsModel,
   useStickyCellStyles,
-} from '../../../lib/components/table/use-sticky-columns';
-import { getStickyClassNames } from '../utils';
-import { renderHook } from '../../__tests__/render-hook';
+} from '../../../../../../lib/components/internal/components/table-fragments/sticky-columns';
+import { getStickyClassNames } from '../../../../../table/utils';
+import { renderHook } from '../../../../../__tests__/render-hook';
 
 function createElementWithWidth(tag: string, width = 0) {
   const element = document.createElement(tag);
