+++ src/internal/components/table-fragments/sticky-columns/index.ts
@@ -1,0 +1,10 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export {
+  selectionColumnId,
+  useStickyColumns,
+  useStickyCellStyles,
+  StickyColumnsModel,
+  StickyColumnsCellState,
+} from './use-sticky-columns';
