+++ src/internal/components/table-fragments/sticky-scrollbar/index.tsx
@@ -1,0 +1,4 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { default as StickyScrollbar } from './sticky-scrollbar';
