+++ src/internal/components/table-fragments/sticky-scrollbar/sticky-scrollbar.tsx
@@ -1,9 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import React, { forwardRef } from 'react';
-import { useAppLayoutContext } from '../internal/context/app-layout-context';
-import { useMergeRefs } from '../internal/hooks/use-merge-refs';
 import { useStickyScrollbar } from './use-sticky-scrollbar';
+import { useMergeRefs } from '../../../hooks/use-merge-refs';
+import { useAppLayoutContext } from '../../../context/app-layout-context';
 import styles from './styles.css.js';
 
 interface StickyScrollbarProps {
