+++ src/table/__integ__/sticky-scrollbar.test.ts
@@ -4,9 +4,9 @@ import useBrowser from '@cloudscape-design/browser-test-tools/use-browser';
 import createWrapper from '../../../lib/components/test-utils/selectors';
 import { BasePageObject } from '@cloudscape-design/browser-test-tools/page-objects';
 const tableWrapper = createWrapper().findTable();
-import styles from '../../../lib/components/table/styles.selectors.js';
+import scrollbarStyles from '../../../lib/components/internal/components/table-fragments/sticky-scrollbar/styles.selectors.js';
 
-const scrollbarSelector = `.${styles['sticky-scrollbar-visible']}`;
+const scrollbarSelector = `.${scrollbarStyles['sticky-scrollbar-visible']}`;
 
 class StickyScrollbarPage extends BasePageObject {
   findVisibleScrollbar() {
