+++ src/table/__tests__/header-cell.test.tsx
@@ -8,7 +8,7 @@ import { renderHook } from '../../__tests__/render-hook';
 
 import styles from '../../../lib/components/table/header-cell/styles.css.js';
 import resizerStyles from '../../../lib/components/table/resizer/styles.css.js';
-import { useStickyColumns } from '../../../lib/components/table/use-sticky-columns';
+import { useStickyColumns } from '../../../lib/components/internal/components/table-fragments/sticky-columns';
 import TestI18nProvider from '../../../lib/components/internal/i18n/testing';
 
 const testItem = {
