+++ src/table/body-cell/td-element.tsx
@@ -3,8 +3,11 @@
 import clsx from 'clsx';
 import React from 'react';
 import styles from './styles.css.js';
-import { StickyColumnsModel, useStickyCellStyles } from '../use-sticky-columns';
 import { getStickyClassNames } from '../utils.js';
+import {
+  StickyColumnsModel,
+  useStickyCellStyles,
+} from '../../internal/components/table-fragments/sticky-columns/use-sticky-columns.js';
 
 export interface TableTdElementProps {
   className?: string;
