+++ src/table/header-cell/index.tsx
@@ -10,9 +10,9 @@ import styles from './styles.css.js';
 import { Resizer } from '../resizer';
 import { useUniqueId } from '../../internal/hooks/use-unique-id';
 import { InteractiveComponent } from '../thead';
-import { StickyColumnsModel, useStickyCellStyles } from '../use-sticky-columns';
 import { getStickyClassNames } from '../utils';
 import { useInternalI18n } from '../../internal/i18n/context';
+import { StickyColumnsModel, useStickyCellStyles } from '../../internal/components/table-fragments/sticky-columns';
 
 interface TableHeaderCellProps<ItemType> {
   className?: string;
