+++ src/table/styles.scss
@@ -149,28 +149,6 @@ filter search icon.
   }
 }
 
-.sticky-scrollbar {
-  height: 15px;
-  position: sticky;
-  display: none;
-  overflow-x: auto;
-  overflow-y: hidden;
-  bottom: 0;
-  width: 100%;
-
-  &-content {
-    height: 15px;
-  }
-  &-visible {
-    display: block;
-  }
-  &-native-invisible {
-    // Native scrollbar does not take space when it is not always invisible
-    // For example it is set as "show when scrolling"
-    margin-top: -15px;
-  }
-}
-
 /*
 The dynamic height dark header needs a background that will cover
 the default white background of the container component.
