+++ src/table/thead.tsx
@@ -10,11 +10,15 @@ import { getColumnKey, getStickyClassNames } from './utils';
 import { TableHeaderCell } from './header-cell';
 import { useColumnWidths } from './use-column-widths';
 import { useVisualRefresh } from '../internal/hooks/use-visual-mode';
-import { selectionColumnId, StickyColumnsModel, useStickyCellStyles } from './use-sticky-columns';
 import styles from './styles.css.js';
 import cellStyles from './header-cell/styles.css.js';
 import headerCellStyles from './header-cell/styles.css.js';
 import ScreenreaderOnly from '../internal/components/screenreader-only';
+import {
+  selectionColumnId,
+  StickyColumnsModel,
+  useStickyCellStyles,
+} from '../internal/components/table-fragments/sticky-columns';
 
 export type InteractiveComponent =
   | { type: 'selection' }
